revamp UI
